Arrange files according to the file structure
