+++ src/classBased/components/TodosList.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import propTypes from 'prop-types';
-import './App.css';
+import '../App.css';
 import TodoItem from './TodoItem';
 
 class TodosList extends Component {
