+++ src/functionBased/components/TodoList.js
@@ -1,0 +1,13 @@
+import Todoitem from './Todoitem';
+
+const TodoList = (props) => {
+  const { todos } = props;
+
+  return (
+    todos.map((task) => (
+      <Todoitem key={task.id} task={task} />
+    ))
+  );
+};
+
+export default TodoList;
